+++ BoothDownloader/src/_Config/Config.cs
