+++ BoothDownloader/src/_Config/JsonConfig.cs
