+++ BoothDownloader/src/_Misc/ChildProgressBarProgress.cs
