+++ BoothDownloader/src/_Web/BoothClient.cs
