+++ BoothDownloader/src/_Web/BoothItem.cs
@@ -1,0 +1,4 @@
+﻿public class BoothItem
+{
+    public string Id = "";
+}
