--- BoothDownloader/src/misc/CancellationTokenValueSource.cs
@@ -1,18 +1,0 @@
-﻿using System.CommandLine.Binding;
-
-namespace BoothDownloader.src.misc
-{
-    public class CancellationTokenValueSource : IValueDescriptor<CancellationToken>, IValueSource
-    {
-        public bool TryGetValue(IValueDescriptor valueDescriptor, BindingContext bindingContext, out object? boundValue)
-        {
-            boundValue = (CancellationToken)bindingContext.GetService(typeof(CancellationToken))!;
-            return true;
-        }
-
-        public object? GetDefaultValue() => throw new NotImplementedException();
-        public string ValueName => throw new NotImplementedException();
-        public Type ValueType => throw new NotImplementedException();
-        public bool HasDefaultValue => throw new NotImplementedException();
-    }
-}
